+++ src/Areas/AppConfig/Commands/KeyValue/KeyValueSetCommand.cs
@@ -13,6 +13,7 @@ public sealed class KeyValueSetCommand(ILogger<KeyValueSetCommand> logger) : Bas
 {
     private const string CommandTitle = "Set App Configuration Key-Value Setting";
     private readonly Option<string> _valueOption = AppConfigOptionDefinitions.Value;
+    private readonly Option<List<string>> _tagsOption = AppConfigOptionDefinitions.Tags;
     private readonly ILogger<KeyValueSetCommand> _logger = logger;
 
     public override string Name => "set";
@@ -21,7 +22,8 @@ public sealed class KeyValueSetCommand(ILogger<KeyValueSetCommand> logger) : Bas
         """
         Set a key-value setting in an App Configuration store. This command creates or updates a key-value setting
         with the specified value. You must specify an account name, key, and value. Optionally, you can specify a
-        label otherwise the default label will be used.
+        label otherwise the default label will be used. You can also specify a content type to indicate how the value
+        should be interpreted. You can add tags in the format 'key=value' to associate metadata with the setting.
         """;
 
     public override string Title => CommandTitle;
@@ -30,12 +32,14 @@ public sealed class KeyValueSetCommand(ILogger<KeyValueSetCommand> logger) : Bas
     {
         base.RegisterOptions(command);
         command.AddOption(_valueOption);
+        command.AddOption(_tagsOption);
     }
 
     protected override KeyValueSetOptions BindOptions(ParseResult parseResult)
     {
         var options = base.BindOptions(parseResult);
         options.Value = parseResult.GetValueForOption(_valueOption);
+        options.Tags = parseResult.GetValueForOption(_tagsOption) ?? [];
         return options;
     }
 
@@ -59,10 +63,11 @@ public sealed class KeyValueSetCommand(ILogger<KeyValueSetCommand> logger) : Bas
                 options.Subscription!,
                 options.Tenant,
                 options.RetryPolicy,
-                options.Label);
-
+                options.Label,
+                options.ContentType,
+                options.Tags);
             context.Response.Results = ResponseResult.Create(
-                new KeyValueSetCommandResult(options.Key, options.Value, options.Label),
+                new KeyValueSetCommandResult(options.Key, options.Value, options.Label, options.ContentType, options.Tags),
                 AppConfigJsonContext.Default.KeyValueSetCommandResult
             );
         }
@@ -75,5 +80,5 @@ public sealed class KeyValueSetCommand(ILogger<KeyValueSetCommand> logger) : Bas
         return context.Response;
     }
 
-    internal record KeyValueSetCommandResult(string? Key, string? Value, string? Label);
+    internal record KeyValueSetCommandResult(string? Key, string? Value, string? Label, string? ContentType = null, List<string>? Tags = null);
 }