+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- App Configuration Put command now also supports setting the `ContentType` and `Tags` properties [#459](https://github.com/Azure/azure-mcp/pull/459)
+
 ### Breaking Changes
 
 ### Bugs Fixed